+++ Changelog.md
@@ -8,8 +8,10 @@
 * Hotfix to the numpy version requirements
 * Changing default grid in all SHT routines to `equiangular`, which makes it consistent with DISCO convolutions
 * Cleaning up the SFNO example and adding new Local Spherical Neural Operator model
+* New filter basis normalization in DISCO convolutions
 * Reworked DISCO filter basis datastructure
 * Support for new filter basis types
+* Adding Morlet-like basis functions on a spherical disk
 
 ### v0.7.2